+++ docs/ReleaseNotes.rst
@@ -6,8 +6,11 @@ Release Notes
 1.5.1
 -----
 Bug fix release:
+
 * MyPolygonSelector.__init__(): define self._polygon_complete.
 * Viewer.grid_from_roi(): Set input_order='none' when follow is False.
+* Series.tags: Ensure that tag lists are always Numpy arrays.
+* Series: Handle properly the situation where a template has wrong number of slices or tags.
 
 1.5.0
 -----